requirements.txt
numpy>=1.21.0
networkx>=2.6.0
plotly>=5.0.0
dash>=2.0.0
dash-bootstrap-components>=1.0.0
litellm>=1.0.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
tqdm>=4.62.0
python-dotenv>=0.19.0 